requirements.txt
Django>=4.2.7
PyJWT>=2.8.0